Initial commit.
In this commit the application can take in the username and password,
with the various parameters needed to search the SAR and MSI data that
is available and return the results.

The application can detect issues with the from date being ahead of the
to date and the to date not being over the current date.